Issue #24
Added a joint search argument in the pubmed hit counts function that defaults to False. If a true argument is passed then it will retrieve the joint hit count as well as the marginals returned as a list. If two MeSH terms are passed to the google distance function then it will utilize this joint search functionality when calling the pubmed hit counts function.