#181
Added code that will run through and create indexes/constraints for several properties on all nodes and then on the rtx_name, id, and name properties for each individual node label. If a new node label is added simply add the new label to node_label_list. (currently on line 523)