Update kg2rollout.md
fixing what I think (but am not 100% certain) is an error in the roll-out template; I think when running the PloverDB `run.sh` script, we need to specify the KG2 release-specific branch of the PloverDB code that we want to be used for building the PloverDB Docker image and container. In most cases, this is a branch called `kg2.X.Yx` where X is the minor version number and Y is the patch release number of RTX-KG2c.